requirements.txt
torch
torchvision
tqdm
pycocotools
Pillow
